[Chore] #2 Review 엔티티 수정
