--- src/main/java/dev/umc/whereseat/domain/stadium/Stadium.java
@@ -1,4 +1,0 @@
-package dev.umc.whereseat.domain.stadium;
-
-public class Stadium {
-}
