+++ src/main/java/dev/umc/whereseat/domain/stadium/entity/Stadium.java
@@ -1,0 +1,4 @@
+package dev.umc.whereseat.domain.stadium.entity;
+
+public class Stadium {
+}
